Bug 1679751 - Move the slab allocator code out of texture_cache.rs. r=gw
A bit of cleanup and a step towards having more allocation algorithms in the texture cache. This patch mostly moves code around, and should not change the behavior of the code.

Depends on D98201

Differential Revision: https://phabricator.services.mozilla.com/D98202
